refactor(backend): restructured backend folder structure
